Moved caching primitives into dedicated partial files
